pyproject.toml
[tool.poetry]
name = "fastapi-template"
version = "1.2.0"

authors = ["Donny Peeters <donny.peeters@hotmail.com>"]
maintainers = ["Donny Peeters <donny.peeters@hotmail.com>"]
description = "A FastAPI template."
readme = "README.md"
homepage = "https://github.com/BiteStreams/fastapi-template"
repository = "https://github.com/BiteStreams/fastapi-template"
keywords = ["FastAPI", "Template"]


[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.31.0"
fastapi = "^0.109.1"
uvicorn = "^0.18.3"
SQLAlchemy = "^1.4.42"
alembic = "^1.8.1"
psycopg2-binary = "^2.9.9"
httpx = "^0.26.0"
redis = "^5.0.1"
pydantic-settings = "^2.2.0"
python-multipart = "^0.0.9"
passlib = "^1.7.4"
python-jose = "^3.3.0"
bcrypt = "^4.1.2"
cloudinary = "^1.39.0"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
black = "^23.12.1"
isort = "^5.13.2"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
markers = [
    "unit: mark a test as a unit test.",
    "integration: mark test as an integration test.",
]

[tool.black]
line_length = 120

[tool.isort]
line_length = 120
profile = "black"
known_first_party = ["api", "infrastructure", "domain"]

[tool.mypy]
plugins = "sqlalchemy.ext.mypy.plugin"
ignore_missing_imports = true
exclude = [".venv"]

